+++ HISTORY.rst
@@ -9,9 +9,12 @@ New features and enhancements
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 * Automatic load of translations on import and possibility to pass translations for virtual modules.
 * New ``xclim.testing.list_datasets`` function listing all available test datasets in repo `xclim-testdata`.
-* New ``sdba.construct_moving_yearly_window`` and ``sdba.unpack_moving_yearly_window`` for moving window adjustments.
 * `spatial_analogs` accepts multi-indexes as the `dist_dim` parameter and will work with candidates and target arrays of different lengths.
 * `humidex` can be computed using relative humidity instead of dewpoint temperature.
+* New ``sdba.construct_moving_yearly_window`` and ``sdba.unpack_moving_yearly_window`` for moving window adjustments.
+* New ``sdba.adjustment.NpdfTransform`` which is an adaptation of Cannon's version of Pitié's _N-dimensional probability density function transform_. Uses new ``sdba.utils.rand_rot_matrix``.
+* New ``sdba.processing.standardize``, ``.unstandardize`` and  ``.reordering``. All of them, tools needed to replicate Cannon's MBCn algorithm.
+* New ``sdba.processing.escore``, backed by  ``sdba.nbutils._escore`` to evaluate the performance of the N pdf transform.
 * New function `xclim.indices.clausius_clapeyron_scaled_precipitation` can be used to scale precipitation according to changes in mean temperature.
 * Percentile based indices gained a `bootstrap` argument that applies a bootstrapping algorithm to reduce biases on exceedance frequencies computed over *in base* and *out of base* periods.
 * Added a `.zenodo.json` file for collecting and maintaining author order and tracking ORCIDs.
@@ -51,6 +54,7 @@ Internal Changes
 * Minor modifications to the GitHub Pull Requests template.
 * Simplification of some yaml elements for virtual modules.
 * Allow injecting `freq` without the missing checks failing.
+
 
 0.27.0 (2021-05-28)
 -------------------